+++ internal/applications/messenger/internal/handler/incoming_push_test.go
@@ -198,7 +198,7 @@ func (p *tsHLSClient) SendRequest(context.Context, string, request.IRequest) err
 }
 
 func (p *tsHLSClient) FetchRequest(context.Context, string, request.IRequest) (response.IResponse, error) {
-	return response.NewResponse().WithCode(200), nil
+	return response.NewResponseBuilder().WithCode(200).Build(), nil
 }
 
 type tsDatabase struct {