+++ pkg/response/types.go
@@ -2,12 +2,16 @@ package response
 
 import "github.com/number571/go-peer/pkg/types"
 
+type IResponseBuilder interface {
+	Build() IResponse
+
+	WithCode(pCode int) IResponseBuilder
+	WithHead(map[string]string) IResponseBuilder
+	WithBody(pBody []byte) IResponseBuilder
+}
+
 type IResponse interface {
 	types.IConverter
-
-	WithCode(pCode int) IResponse
-	WithHead(map[string]string) IResponse
-	WithBody(pBody []byte) IResponse
 
 	GetCode() int
 	GetHead() map[string]string